add fibonacci evaluation

+++ app/src/main/java/ru/ppr/pprtest/adapters/NumberItem.kt
@@ -1,6 +1,8 @@
 package ru.ppr.pprtest.adapters
 
+import java.math.BigInteger
+
 data class NumberItem(
-    val number: Int,
+    val number: BigInteger,
     val isWhiteColor: Boolean = true
 )
